+++ projects/forge/src/main/kotlin/site/siredvin/peripheralworks/integrations/fluxnetworks/Configuration.kt
@@ -8,13 +8,13 @@ object Configuration : IForgeConfigHandler {
     private var enableFluxControllerConfig: ForgeConfigSpec.BooleanValue? = null
 
     val enableFluxController: Boolean
-        get() = enableFluxControllerConfig?.get() ?: true
+        get() = enableFluxControllerConfig?.get() != false
 
     override val name: String
         get() = "flux_networks"
 
     override fun addToConfig(builder: ForgeConfigSpec.Builder) {
         enableFluxControllerConfig = builder.comment("Enables flux controller integration")
-            .define("enableAutomobile", true)
+            .define("enableFluxController", true)
     }
 }